refactor(core): decouple orchestration from app.remote
