+++ app/cli/ui/renderer/__init__.py
@@ -3,6 +3,5 @@
 from __future__ import annotations
 
 from app.cli.ui.renderer.renderer import StreamRenderer, _canonical_node_name
-from app.remote.stream import StreamEvent
 
-__all__ = ["StreamEvent", "StreamRenderer", "_canonical_node_name"]
+__all__ = ["StreamRenderer", "_canonical_node_name"]
