+++ app/remote/__init__.py
@@ -3,6 +3,5 @@
 from __future__ import annotations
 
 from app.remote.client import RemoteAgentClient, RemoteRunResult
-from app.remote.stream import StreamEvent
 
-__all__ = ["RemoteAgentClient", "RemoteRunResult", "StreamEvent"]
+__all__ = ["RemoteAgentClient", "RemoteRunResult"]
